+++ docs/changelog.rst
@@ -24,6 +24,8 @@ Changelog
   multithreaded database access could cause an internal error (with the message
   "database is locked"). This release synchronizes access to the database to
   avoid internal SQLite contention, which should avoid this error.
+* Plugins can now add parallel stages to the import pipeline. See
+  :ref:`writing-plugins`.
 * New plugin event: ``import_task_choice`` is called after an import task has an
   action assigned.
 * New plugin event: ``library_opened`` is called when beets starts up and